+++ src/pwshBedrock/Private/Format-MetaTextMessage.ps1
@@ -131,7 +131,9 @@ function Format-MetaTextMessage {
             'meta.llama3-2-3b-instruct-v1:0',
             'meta.llama3-2-11b-instruct-v1:0',
             'meta.llama3-2-90b-instruct-v1:0',
-            'meta.llama3-3-70b-instruct-v1:0'
+            'meta.llama3-3-70b-instruct-v1:0',
+            'meta.llama4-maverick-17b-instruct-v1:0',
+            'meta.llama4-scout-17b-instruct-v1:0'
         )]
         [string]$ModelID,