+++ src/pwshBedrock/Public/Reset-ModelContext.ps1
@@ -79,6 +79,8 @@ function Reset-ModelContext {
             'meta.llama3-2-11b-instruct-v1:0',
             'meta.llama3-2-90b-instruct-v1:0',
             'meta.llama3-3-70b-instruct-v1:0',
+            'meta.llama4-maverick-17b-instruct-v1:0',
+            'meta.llama4-scout-17b-instruct-v1:0',
             'mistral.mistral-7b-instruct-v0:2',
             'mistral.mistral-large-2402-v1:0',
             'mistral.mistral-large-2407-v1:0',
@@ -143,6 +145,8 @@ function Reset-ModelContext {
                         $model -eq 'meta.llama3-2-11b-instruct-v1:0' -or
                         $model -eq 'meta.llama3-2-90b-instruct-v1:0' -or
                         $model -eq 'meta.llama3-3-70b-instruct-v1:0' -or
+                        $model -eq 'meta.llama4-maverick-17b-instruct-v1:0' -or
+                        $model -eq 'meta.llama4-scout-17b-instruct-v1:0' -or
                         $model -eq 'mistral.mistral-7b-instruct-v0:2' -or
                         $model -eq 'mistral.mixtral-8x7b-instruct-v0:1' -or
                         $model -eq 'mistral.mistral-large-2402-v1:0' -or
@@ -185,6 +189,8 @@ function Reset-ModelContext {
                             $model -eq 'meta.llama3-2-11b-instruct-v1:0' -or
                             $model -eq 'meta.llama3-2-90b-instruct-v1:0' -or
                             $model -eq 'meta.llama3-3-70b-instruct-v1:0' -or
+                            $model -eq 'meta.llama4-maverick-17b-instruct-v1:0' -or
+                            $model -eq 'meta.llama4-scout-17b-instruct-v1:0' -or
                             $model -eq 'mistral.mistral-7b-instruct-v0:2' -or
                             $model -eq 'mistral.mixtral-8x7b-instruct-v0:1' -or
                             $model -eq 'mistral.mistral-large-2402-v1:0' -or